Peterk/couchbase (#45)
* added couchbase connector

The couchbase connector is a struct which is called a client. Each client can retrieve data based on the clientID.

* updated couchbase connector

* Updated the couchbase connector

It has been generalized so that only one will be needed to do all couchbase data interactions.

* added couchbase connector

The couchbase connector is a struct which is called a client. Each client can retrieve data based on the clientID.

* updated couchbase connector

* Updated the couchbase connector

It has been generalized so that only one will be needed to do all couchbase data interactions.

* Started testing the couchbase connector.

* updated repo structure

* added couchbase connector

The couchbase connector is a struct which is called a client. Each client can retrieve data based on the clientID.

* updated couchbase connector

* Updated the couchbase connector

It has been generalized so that only one will be needed to do all couchbase data interactions.

* added couchbase connector

The couchbase connector is a struct which is called a client. Each client can retrieve data based on the clientID.

* Started testing the couchbase connector.

* Updated the api

The API now only accepts clients who have a document that has been made for them. The API also automatically adds the eventID to the couchbase document if it is not already there.