Peterk/cb config (#47)
* added couchbase connector

The couchbase connector is a struct which is called a client. Each client can retrieve data based on the clientID.

* updated couchbase connector

* Updated the couchbase connector

It has been generalized so that only one will be needed to do all couchbase data interactions.

* added couchbase connector

The couchbase connector is a struct which is called a client. Each client can retrieve data based on the clientID.

* updated couchbase connector

* Updated the couchbase connector

It has been generalized so that only one will be needed to do all couchbase data interactions.

* Started testing the couchbase connector.

* updated repo structure

* added couchbase connector

The couchbase connector is a struct which is called a client. Each client can retrieve data based on the clientID.

* updated couchbase connector

* Updated the couchbase connector

It has been generalized so that only one will be needed to do all couchbase data interactions.

* added couchbase connector

The couchbase connector is a struct which is called a client. Each client can retrieve data based on the clientID.

* Started testing the couchbase connector.

* Updated the api

The API now only accepts clients who have a document that has been made for them. The API also automatically adds the eventID to the couchbase document if it is not already there.

* added 'config' variable for couchbase

* add couchbase 'config' variables

* Added config settings for couchbase

Updated the readme with instructions on couchbase setup

* added couchbase connector

The couchbase connector is a struct which is called a client. Each client can retrieve data based on the clientID.

* updated couchbase connector

* Updated the couchbase connector

It has been generalized so that only one will be needed to do all couchbase data interactions.

* added couchbase connector

The couchbase connector is a struct which is called a client. Each client can retrieve data based on the clientID.

* Started testing the couchbase connector.

* added couchbase connector

The couchbase connector is a struct which is called a client. Each client can retrieve data based on the clientID.

* added couchbase connector

The couchbase connector is a struct which is called a client. Each client can retrieve data based on the clientID.

* updated couchbase connector

* Updated the couchbase connector

It has been generalized so that only one will be needed to do all couchbase data interactions.

* Started testing the couchbase connector.

* Updated the api

The API now only accepts clients who have a document that has been made for them. The API also automatically adds the eventID to the couchbase document if it is not already there.

* added 'config' variable for couchbase

* add couchbase 'config' variables

* Added config settings for couchbase

Updated the readme with instructions on couchbase setup

* fixed merge conflicts

* Updated config setup

The config variable is one connection string and several scripts and folder have been renamed to follow camle case rather than use _